+++ src/fields/Months.tsx
@@ -30,11 +30,6 @@ export default function Months(props: MonthsProps) {
     [className]
   )
 
-  const units = {
-    ...UNITS[3],
-    alt: locale.monthsAlt || UNITS[3].alt,
-  } as Unit
-
   return (
     <div className={internalClassName}>
       {locale.prefixMonths !== '' && (
@@ -46,7 +41,12 @@ export default function Months(props: MonthsProps) {
         optionsList={optionsList}
         grid={false}
         value={value}
-        unit={units}
+        unit={{
+          ...UNITS[3],
+          // Allow translation of alternative labels when using "humanizeLabels"
+          // Issue #3
+          alt: locale.altMonths || DEFAULT_LOCALE_EN.altMonths,
+        }}
         setValue={setValue}
         locale={locale}
         className={className}